Add `<NextPlainImage />` component
This component is based on the Next.js `<Image />` component, but it's
extremely stripped back. In my JSON Feed testing, I found that NetNewsWire
doesn't support the `<noscript>` element. This means that the Next.js
`<Image />` component doesn't work.

`<NextPlainImage />` contains only the basic functionality required to
render a Next.js image in a feed.